# go/model_leaderboard.go
/*
 * Leaderboard API
 *
 * Leaderaboard API
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Leaderboard struct {

	Uuid string `json:"uuid"`
}