# go/model_score.go
/*
 * Leaderboard API
 *
 * Leaderaboard API
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Score struct {

	EntityId string `json:"entityId"`

	Score float64 `json:"score"`
}